requirements.txt
fastapi
uvicorn
pillow
pillow-heif
python-dotenv
requests